Simplify parser module
